# backend/auth_service/openapi.yaml
openapi: 3.0.0
info:
  title: CandidateV Authentication Service API
  version: 1.0.0
  description: API for user authentication and token management
  contact:
    name: CandidateV Support
    email: support@candidatev.com
servers:
  - url: http://localhost:8000
    description: Local development server
  - url: https://{auth_service_url}
    description: Production server
    variables:
      auth_service_url:
        default: auth-service.candidatev.com
components:
  schemas:
...
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          description: User's email address
          example: john.doe@example.com
        password:
          type: string
          format: password
          description: User's password
          example: securepassword123
    TokenResponse:
      type: object
      required:
        - access_token
        - refresh_token
        - token_type
        - expires_in
      properties:
        access_token:
          type: string
          description: JWT access token
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        refresh_token:
          type: string
          description: JWT refresh token
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        token_type:
          type: string
          enum: [bearer]
          description: Type of token
          example: bearer
        expires_in:
          type: integer
          description: Token expiration in seconds
          example: 1800
...
    ErrorResponse:
      type: object
      required:
        - status
        - message
      properties:
        status:
          type: string
          enum: [error]
          description: Error status
          example: error
        message:
          type: string
          description: Error message
          example: Invalid credentials
        code:
          type: string
          description: Error code
          example: INVALID_CREDENTIALS
        requestId:
          type: string
          description: Request ID for tracing
          example: 123e4567-e89b-12d3-a456-426614174000
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token authentication
paths:
...
  /api/auth/login:
    post:
      summary: User login
      description: Authenticate user and get access token
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...